pyproject.toml
[project]
name = "flow-fields"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.6",
    "pycairo>=1.28.0",
    "tqdm>=4.67.1",
]
